+++ client/src/components/Header.css
@@ -1,0 +1,17 @@
+.header {
+  padding-left: 10%;
+  padding-right: 10%;
+  padding-top: 3%;
+  padding-bottom: 3%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: rgb(3, 30, 74);
+  color: white;
+}
+
+.language-container img {
+  width: 45px;
+  padding: 10px;
+}
