+++ client/src/components/footer.css
@@ -1,0 +1,40 @@
+body {
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu,
+    Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  padding: 0;
+  margin: 0;
+}
+.main-container {
+  padding-left: 10%;
+  padding-right: 10%;
+  padding-top: 3%;
+  padding-bottom: 3%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: rgb(3, 30, 74);
+  color: white;
+}
+
+.main-container li {
+  display: flex;
+  gap: 10px;
+}
+
+.main-container img {
+  filter: invert(100%) sepia(0%) saturate(1681%) hue-rotate(312deg) brightness(102%) contrast(102%);
+}
+
+.icon-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.icon-container img {
+  width: 45px;
+  padding: 10px;
+  filter: invert(100%) sepia(0%) saturate(1681%) hue-rotate(312deg) brightness(102%) contrast(102%);
+}
