+++ client/src/constants.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:3000/api/';
